# core/openapi.yaml
openapi: 3.1.0
info:
  title: Madang core API
  version: 2.0.0
  summary: madang-core의 로컬 HTTP·WebSocket API.
  description: |
    데스크톱 앱은 이 API로 madang-core와 통신한다. 페이지 기록은 프로젝트
    폴더의 `.madang/`에, 전역 설정(config.yaml)과 profile.md는 앱 홈(`~/.madang`)에 있다.
    둘 다 core가 쓰며 앱은 자체 문서를 두지 않는다. core는 기록을 git에
    커밋하지 않는다(커밋 여부는 사용자가 정한다).

    규약:
    - 시각은 오프셋이 붙은 ISO 8601이다(`2026-09-24T09:40:00+09:00`).
    - 페이지 id는 `YYYY-MM-DD-<slug>`, 블록 id는 `bNN`이다. run 번호는
      페이지별로 늘어나며 재사용하지 않는다.
    - 페이지 안의 경로는 페이지 기준이다(`blocks/b05-base.json`). 프로젝트
      폴더 기준 경로에는 `repo:` 접두어가 붙는다
      (`repo:src/session/lock.ts`).
    - 오류는 `ApiError`로 돌려준다. 파일 검사에 실패하면 400과 함께
      `ValidationFailure`와 그 `issues`를 돌려준다.
    - 변경은 `/events` WebSocket으로 알린다. 앱은 재연결 후 전부 다시
      가져온다.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:7470
    description: 로컬 core. 포트는 다를 수 있으며 core가 core.port에 적는다.
...
paths:
...
  /projects/{project}/git/status:
    parameters:
      - $ref: "#/components/parameters/Project"
    get:
      tags: [git]
      operationId: getGitStatus
      summary: 작업 트리의 변경 파일, 브랜치, HEAD.
      description: |
        git 저장소가 아니면 `repository`가 거짓이고 앱은 "git 시작"
        (`POST /projects/{project}/git/init`)만 보여 준다. 앱 밖에서 한 git
        조작은 여기서 상태로만 보인다.
      parameters:
        - $ref: "#/components/parameters/PageQuery"
      responses:
        "200":
          description: 상태.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/diff:
    parameters:
      - $ref: "#/components/parameters/Project"
    get:
      tags: [git]
      operationId: getGitDiff
      summary: 통합 diff 텍스트(디프 탭).
      description: |
        기본은 작업 트리와 인덱스의 차이다. `staged`면 인덱스와 HEAD,
        `base`·`target`을 주면 두 커밋의 차이다.
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - name: staged
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: base
          in: query
          required: false
          description: 비교 기준 커밋이나 브랜치.
          schema:
            type: string
        - name: target
          in: query
          required: false
          description: 비교 대상 커밋이나 브랜치.
          schema:
            type: string
        - name: path
          in: query
          required: false
          description: 이 경로들로 제한한다. 반복할 수 있다.
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: diff.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitDiff"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/stage:
    parameters:
      - $ref: "#/components/parameters/Project"
    post:
      tags: [git]
      operationId: stageGit
      summary: 변경을 스테이징한다.
      description: "`paths`가 없으면 모든 변경이다. 끝나면 `git.changed`를 낸다."
      parameters:
        - $ref: "#/components/parameters/PageQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitStage"
      responses:
        "200":
          description: 스테이징 뒤의 상태.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/commit:
    parameters:
      - $ref: "#/components/parameters/Project"
    post:
      tags: [git]
      operationId: commitGit
      summary: 스테이징된 변경을 커밋한다.
      description: |
        `paths`를 주면 그 경로만 커밋한다. 스테이징된 것이 없거나 비밀로
        보이는 파일(`.env`, `*.pem` 등)이 바뀌어 있으면 409. 정책이 막으면
        409 `denied`. 끝나면 `git.changed`를 낸다.
      parameters:
        - $ref: "#/components/parameters/PageQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitCommitCreate"
      responses:
        "200":
          description: 새 커밋.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoCommitResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/branches:
    parameters:
      - $ref: "#/components/parameters/Project"
    get:
      tags: [git]
      operationId: listGitBranches
      summary: 로컬 브랜치와 메인 체크아웃의 현재 브랜치.
      responses:
        "200":
          description: 브랜치.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitBranches"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/worktrees:
    parameters:
      - $ref: "#/components/parameters/Project"
    get:
      tags: [git]
      operationId: listGitWorktrees
      summary: 메인 체크아웃과 워크트리, 그 워크트리를 쓰는 페이지.
      description: |
        코드 페이지의 워크트리는 `<프로젝트>.wt/<page-id>/`에 있고 `page`에
        그 페이지 id가 있다. 메인 체크아웃은 `main`이 참이다.
      responses:
        "200":
          description: 워크트리 목록.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitWorktree"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/log:
    parameters:
      - $ref: "#/components/parameters/Project"
    get:
      tags: [git]
      operationId: getGitLog
      summary: "`ref`에서 거슬러 올라가는 커밋, 최근 것부터."
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - name: ref
          in: query
          required: false
          schema:
            type: string
            default: HEAD
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: 커밋 목록.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitLogEntry"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/push:
    parameters:
      - $ref: "#/components/parameters/Project"
    post:
      tags: [git]
      operationId: pushGit
      summary: 현재 브랜치를 같은 이름의 원격 브랜치로 강제 없이 보낸다.
      description: |
        브랜치에 설정된 원격, 없으면 `origin`으로 보낸다. 정책이 막으면 409
        `denied`, 원격이 없거나 거부되면 409. 끝나면 `git.changed`를 낸다.
      parameters:
        - $ref: "#/components/parameters/PageQuery"
      responses:
        "200":
          description: push했다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoPushResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/pull:
    parameters:
      - $ref: "#/components/parameters/Project"
    post:
      tags: [git]
      operationId: pullGit
      summary: 빨리 감기로만 당겨 온다.
      description: 갈라졌거나 원격이 없으면 409. 끝나면 `git.changed`를 낸다.
      parameters:
        - $ref: "#/components/parameters/PageQuery"
      responses:
        "200":
          description: 당긴 뒤의 상태.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/git/init:
    parameters:
      - $ref: "#/components/parameters/Project"
    post:
      tags: [git]
      operationId: initGit
      summary: 프로젝트 폴더를 git 저장소로 만든다("git 시작").
      description: |
        `.madang/`은 프로젝트 `config.yaml`의 `track`대로 `.git/info/exclude`에
        더한다. 이미 저장소면 409. 커밋은 만들지 않는다.
      responses:
        "201":
          description: 새 저장소의 상태.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    Project:
      name: project
      in: path
      required: true
      description: 프로젝트 id.
      schema:
        type: string
      example: jobs
...
    PageQuery:
      name: page
      in: query
      required: false
      description: 이 페이지의 워크트리(있으면)에서 일한다. 없으면 프로젝트 폴더.
      schema:
        type: string
      example: 2026-09-24-lock
  responses:
    Invalid:
      description: 요청 또는 결과 파일이 검사에 실패했다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationFailure"
    NotFound:
      description: 프로젝트, 페이지, 블록, run, 파일이 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: not_found
            message: page '2026-09-30-missing' not found
    Conflict:
      description: 현재 상태에서는 변경할 수 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: conflict
            message: "project 'jobs' still has child projects: jobs-2025"
  schemas:
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: >-
            기계가 읽는 사유: `not_found`, `conflict`, `invalid`, `busy`,
            `no_repo`, `denied`(정책의 금지 명령), 게시 오류 코드
            (`target-missing`, `nothing-to-undo`, `target-changed` 등).
        message:
          type: string
      examples:
        - error: conflict
          message: a run is already in progress on this page
        - error: denied
          message: "금지 명령: push --force"
    Issue:
      type: object
      description: 검사 문제 하나.
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            케밥 표기 식별자. 알려진 코드: `missing-file`,
            `unreadable`, `frontmatter`, `missing-key`, `invalid-value`,
            `duplicate-decision`, `choice-not-in-options`, `repo-unset`,
            `invalid-artifact`, `artifact-missing`, `missing-section`,
            `token-limit`, `done-without-verify`, `invalid-yaml`.
        message:
          type: string
        line:
          type: [integer, "null"]
          description: 1부터 세는 파일 줄. 알 때만.
        path:
          type: [string, "null"]
          description: 파일. 알 때만.
      examples:
        - code: missing-section
          message: "missing required section '## 다음 할 일'"
          line: null
          path: ledger.md
    ValidationFailure:
      type: object
      required: [error, message, issues]
      properties:
        error:
          type: string
          enum: [invalid]
        message:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
      examples:
        - error: invalid
          message: ledger.md failed validation
          issues:
            - code: invalid-value
              message: "status: expected planning | doing | blocked | review | done"
              line: 2
              path: ledger.md
            - code: token-limit
              message: ledger.md is 2310 tokens (limit 2000); summarize the log
              line: null
              path: ledger.md

    # 시스템
...
    RepoCommitResult:
      type: object
      required: [commit]
      properties:
        commit:
          type: string
      examples:
        - commit: 7b8c9d0
    RepoPushResult:
      type: object
      required: [branch]
      properties:
        branch:
          type: string
        remote:
          type: string
      examples:
        - branch: main
          remote: origin
...
    GitFile:
      type: object
      description: 바뀐 파일 하나.
      required: [path, code]
      properties:
        path:
          type: string
          description: 작업 트리 기준 경로.
        code:
          type: string
          description: "`git status --porcelain`의 두 글자 코드(인덱스, 작업 트리). 추적 안 되는 파일은 `??`."
      examples:
        - path: src/session/lock.ts
          code: " M"
    GitStatus:
      type: object
      description: 작업 트리 상태. git 저장소가 아니면 `repository`만 거짓이다.
      required: [repository, folder, files]
      properties:
        repository:
          type: boolean
        folder:
          type: string
          description: 작업 트리의 절대 경로(프로젝트 폴더 또는 페이지 워크트리).
        branch:
          type: string
          description: 체크아웃된 브랜치. HEAD가 분리됐으면 없다.
        head:
          type: string
          description: HEAD 커밋의 전체 해시. 커밋이 없으면 없다.
        files:
          type: array
          items:
            $ref: "#/components/schemas/GitFile"
      examples:
        - repository: true
          folder: /Users/me/src/madang.wt/2026-09-24-lock
          branch: page/2026-09-24-lock
          head: 7b8c9d0e1f2a3b4c5d6e7f8091a2b3c4d5e6f708
          files:
            - {path: src/session/lock.ts, code: " M"}
            - {path: src/session/lock.test.ts, code: "??"}
        - repository: false
          folder: /Users/me/Documents/jobs
          files: []
    GitDiff:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: 통합 diff 텍스트(색 없음).
      examples:
        - text: |
            diff --git a/src/session/lock.ts b/src/session/lock.ts
            --- a/src/session/lock.ts
            +++ b/src/session/lock.ts
            @@ -1 +1,2 @@
             export {}
            +export const refreshLock = new Mutex()
    GitStage:
      type: object
      properties:
        paths:
          type: [array, "null"]
          description: 스테이징할 경로. 없으면 모든 변경.
          items:
            type: string
      examples:
        - paths: [src/session/lock.ts]
        - {}
    GitCommitCreate:
      type: object
      required: [message]
      properties:
        message:
          type: string
        paths:
          type: [array, "null"]
          description: 이 경로들만 커밋한다.
          items:
            type: string
      examples:
        - message: "feat(session): add refresh lock"
    GitBranches:
      type: object
      required: [branches]
      properties:
        current:
          type: string
          description: 메인 체크아웃의 브랜치. HEAD가 분리됐으면 없다.
        branches:
          type: array
          items:
            type: string
      examples:
        - current: main
          branches: [main, page/2026-09-24-lock]
    GitWorktree:
      type: object
      required: [path, head, main]
      properties:
        path:
          type: string
        branch:
          type: string
          description: 체크아웃된 브랜치. 분리됐으면 없다.
        head:
          type: string
          description: HEAD 해시. 커밋이 없으면 빈 문자열.
        main:
          type: boolean
          description: 메인 체크아웃이면 참.
        page:
          type: string
          description: 이 워크트리를 쓰는 페이지 id.
      examples:
        - path: /Users/me/src/madang
          branch: main
          head: 7b8c9d0e1f2a3b4c5d6e7f8091a2b3c4d5e6f708
          main: true
        - path: /Users/me/src/madang.wt/2026-09-24-lock
          branch: page/2026-09-24-lock
          head: 1a2b3c4d5e6f708192a3b4c5d6e7f8091a2b3c4d
          main: false
          page: 2026-09-24-lock
    GitLogEntry:
      type: object
      required: [hash, author, date, subject]
      properties:
        hash:
          type: string
        author:
          type: string
        date:
          type: string
          format: date-time
        subject:
          type: string
          description: 메시지 첫 줄.
      examples:
        - hash: 7b8c9d0e1f2a3b4c5d6e7f8091a2b3c4d5e6f708
          author: choi
          date: "2026-09-24T09:40:00+09:00"
          subject: "feat(session): add refresh lock"

    # 되돌리기
...